fix TestIntegrationGetReleaseTargets

+++ clients/cockroachdb/client_test.go
@@ -1044,7 +1044,7 @@ func TestIntegrationGetReleaseTargets(t *testing.T) {
 			assert.Equal(t, "", releaseTargets)
 		}
 		assert.Equal(t, "GetReleaseTargets", releaseTargets[0]["name"])
-		assert.Equal(t, 2, releaseTargets[0]["pipelinescount"])
+		assert.Equal(t, int64(2), releaseTargets[0]["pipelinescount"])
 	})
 }
